// app/types/myTypes.ts
export type WorkProps = {
  image: string;
  title: string;
  role: string;
  date: string;
  url: string;
};

export type SideProjectsProps = {
  title: string;
  description: string;
  url: string;
};

export type ContactMeProps = {
  image: string;
  mediumImage: string;
  name: string;
  description: string;
};